+++ src/core/actions.ts
@@ -217,6 +217,13 @@ export type RootAction =
       }
     }
   | {
+      type: 'CURRENCY_ENGINE_SEEN_TX_CHECKPOINT_CHANGED'
+      payload: {
+        walletId: string
+        checkpoint: string
+      }
+    }
+  | {
       // Called when a currency engine is wiped out.
       type: 'CURRENCY_ENGINE_CLEARED'
       payload: {